Refactor layout structure and styles
