+++ src/assets/images/job-offers-banner.png
