+++ src/layouts/Layout.astro
@@ -1,22 +1,167 @@
-<!doctype html>
-<html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width" />
-		<link rel="icon" type="image/svg+xml" href="/favicon.svg" />
-		<meta name="generator" content={Astro.generator} />
-		<title>Astro Basics</title>
-	</head>
-	<body>
-		<slot />
-	</body>
-</html>
+---
+import "../assets/stylesheets/application.css"
 
-<style>
-	html,
-	body {
-		margin: 0;
-		width: 100%;
-		height: 100%;
-	}
-</style>
+import logoMensa from "../assets/images/logo-mensa-peru.png"
+---
+
+<!DOCTYPE html>
+<html lang="es-PE">
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="icon" href="static/img/favicon-32x32.png" type="image/png">
+    <title>Mensa Perú</title>
+    <link rel="preload" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.7.2/css/all.min.css" as="style">
+    <script async src="https://www.googletagmanager.com/gtag/js?id=G-36NY3N7VE3"></script>
+    <script>
+      window.dataLayer = window.dataLayer || [];
+      function gtag(){dataLayer.push(arguments);}
+      gtag('js', new Date());
+
+      gtag('config', 'G-36NY3N7VE3');
+    </script>
+		<script src="../assets/javascript/application.js"></script>
+  </head>
+  <body data-dark-theme="true">
+    <!-- Navigation Bar -->
+    <nav class="navbar">
+      <div class="overlay"></div>
+
+      <div class="container">
+        <div class="navbar-logo">
+          <img src={logoMensa.src} alt="Logo de Mensa Perú" class="logo">
+          <span class="org-name">Mensa Perú</span>
+        </div>
+
+        <div class="navbar-toggle">
+          <button class="toggle-button" type="button" title="Abrir menú">
+            <i class="fa-solid fa-bars"></i>
+          </button>
+        </div>
+
+        <div class="navbar-content">
+          <div class="theme-toggle">
+            <button class="dark-mode-toggle" type="button" title="Cambiar tema">
+              <i class="fas fa-sun active"></i>
+              <i class="fas fa-moon"></i>
+            </button>
+            <button class="toggle-button" type="button" title="Cerrar menú">
+              <i class="fa-solid fa-times"></i>
+            </button>
+          </div>
+
+          <ul class="navbar-nav">
+            <li class="nav-item dropdown">
+              <span class="dropdown-toggle" role="button">
+                Sobre Nosotros
+                <i class="dropdown-icon fa-solid fa-chevron-down"></i>
+              </span>
+              <ul class="dropdown-menu">
+                <li class="nav-item">
+                  <a class="nav-link" href="#who-we-are">Nosotros</a>
+                </li>
+                <li class="nav-item">
+                  <a class="nav-link" href="#our-mission">Misión</a>
+                </li>
+                <li class="nav-item">
+                  <a class="nav-link" href="consejo-directivo.html">Directiva</a>
+                </li>
+              </ul>
+            </li>
+
+            <li class="nav-item">
+              <a class="nav-link" href="#what-we-offer">¿Qué ofrecemos?</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="#join-us">Únete</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="ofertas-empleo.html">Empleos</a>
+            </li>
+          </ul>
+        </div>
+
+        <div class="theme-toggle">
+          <button class="dark-mode-toggle" type="button" title="Cambiar tema">
+            <i class="fas fa-sun active"></i>
+            <i class="fas fa-moon"></i>
+          </button>
+        </div>
+      </div>
+    </nav>
+
+    <main>
+			<slot />
+		</main>
+
+    <!-- Footer -->
+    <footer class="footer">
+      <div class="container">
+        <div class="links">
+          <div>
+            <a href="https://www.mensa.org/" target="_blank" rel="noopener">
+              Mensa Internacional
+            </a>
+          </div>
+          <div>
+            <a href="mailto:info@mensa.pe">
+              Contacto
+            </a>
+          </div>
+          <div>
+            <a href="https://portal.mensa.pe/" target="_blank" rel="noopener">
+              Acceso Miembros
+            </a>
+          </div>
+        </div>
+
+        <div class="social-media">
+          <div>
+            <a href="https://facebook.com/Mensa.Peru/"
+               target="_blank"
+               rel="noopener"
+               title="Síguenos en Facebook"
+            >
+              <i class="fab fa-facebook fa-2x"></i>
+            </a>
+          </div>
+
+          <div>
+            <a href="https://www.linkedin.com/company/mensa-per%C3%BA/"
+               target="_blank"
+               rel="noopener"
+               title="Síguenos en LinkedIn"
+            >
+              <i class="fab fa-linkedin fa-2x"></i>
+            </a>
+          </div>
+
+          <div>
+            <a href="https://github.com/Mensa-Peru/"
+               target="_blank"
+               rel="noopener"
+               title="Síguenos en GitHub"
+            >
+              <i class="fab fa-github fa-2x"></i>
+            </a>
+          </div>
+        </div>
+      </div>
+
+      <div class="copyrigth">
+        <p>&copy; 2025 Mensa Perú.</p>
+      </div>
+    </footer>
+		<script>
+			function isDarkTheme() {
+				const storedValue = localStorage.getItem("dark-theme")
+				const mediaQuery = window.matchMedia("(prefers-color-scheme: dark)")
+				const isDark = storedValue || mediaQuery.matches
+
+				return isDark
+			}
+
+			document.body.setAttribute('data-dark-theme', isDarkTheme())
+		</script>
+  </body>
+</html>
